requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
conda==4.8.5
conda-package-handling==1.7.0+0.g7c4a471.dirty
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.1.2
django-allow-cidr==0.3.1
django-crispy-forms==1.9.2
djangorestframework==3.12.1
fake-useragent==0.1.11
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
netaddr==0.8.0
numpy @ file:///opt/concourse/worker/volumes/live/c646f6b9-b398-438d-7dc7-c7142e0ba44b/volume/numpy_and_numpy_base_1596235504536/work
parse==1.18.0
parsel==1.6.0
Pillow==8.0.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
PyDispatcher==2.0.5
pyee==7.0.4
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyppeteer==0.2.2
pyquery==1.4.1
PySocks==1.7.1
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
requests-html==0.10.0
ruamel-yaml==0.15.87
Scrapy==2.4.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.46.0
Twisted==20.3.0
urllib3==1.25.8
w3lib==1.22.0
webdriver-manager==3.2.2
websockets==8.1
whitenoise==5.2.0
zope.interface==5.1.2
